aaron_requirements.txt
numpy==1.25.1
InstructorEmbedding==1.0.1
psycopg2-binary==2.9.6
torch==2.0.1
tqdm==4.65.0
sentence_transformers==2.2.2
openai==0.27.8